Remove export:mappings task and code
This removes a task that produces a CSV file for transition to import
each day.

This task had been failing for a while as Whitehall no longer has a
`production_slave` db config (presumably since Rails 7 update).

Once I looked into it, I decided that this task was somewhat unnecessary
for GOV.UK Transition for 2022. It is a rather complex process that maps
URLs used as part of Transition to pages in Whitehall, for keeping track
of their destination and redirects. In 2022 most things in Transition
were moved over around 9 years ago and it doesn't look like any of these
Whitehall mappings have had updates since 2016.

This pairs with: https://github.com/alphagov/transition/pull/1165 which
removes the corresponding import task from Transition.