Only queue jobs for attachments that have changed
This commit changes the behaviour of the admin attachments controller so that it only queues up Sidekiq jobs for attachments that have changed.

Prior to this, making a change to one attachment would cause `AssetManagerAttachmentMetadataWorker` workers to be queued up for every attachment in the parent document.

In extreme cases, for example when one document has hundreds of attachments, it meant that a single change to one attachment would result in hundreds of jobs being added to the Sidekiq queue. This had the potential to overwhelm the queue if multiple attachments were changed in quick succession – i.e. quicker than Sidekiq could process all the workers.

Instead, we'll now only queue up workers for the individual attachment that has been changed.