Add an async worker for updating asset attributes
This commit introduces an "update Whitehall asset" worker (`AssetManagerUpdateWhitehallAssetWorker`) which can be used to asynchronously update the attributes associated with a file in Asset Manager. The worker does this by calling the `AssetManager::AssetUpdater` service.

The worker accepts models that have a `file` CarrierWave uploader attribute – i.e. `AttachmentData`, `ImageData` and `ConsultationResponseFormData`.