Remove table header partial
This partial is only used in one view and contains only the table
headers for a table that is rendered elsewhere.

This is confusing and has led to the header and content getting out of
sync (see previous commit). Moving the header out of a partial so it is
easier to work with.