Add government_id and political to the HtmlAttachmentPresenter
This allows us to identify when the edition that the html_publication
belongs to is historically political. When it is we will render a
banner on the frontend which indicates that the attachment was published
under a historical government.

It works passing the political value through as a key value pair on the
description and the government_id as a link.

This link is expanded to include all the government details by Content
Store before being passed to the frontend. More information can be
seen in the govuk-content-schema docs on how this works.