Symlink app/asset-manager-tmp directory to /tmp, just in case.
Getting a bit desperate at this point, but seems worth a try in the
continued absence of any errors or logs relating to what's going wrong.
Whatever's failing has resisted all attempts to reduce it to a small
testcase so far.

Also make the `mkdir /app` fail if `/app` already exists, just so we're
really sure we'll always be created the symlinks where we intended.