Fix test coverage reporting when running with parallelization
When running tests with multiple parallel workers (as in C.I. and locally by default), SimpleCov was not correctly merging the coverage reports from each individual worker. As a result, our test coverage was being mis-reported as 63%.

The issue dates back to commit 979437 which removed the parallel_tests gem, which SimpleCov detects automatically and works well alongside, replacing it with Rails 6's native test parallelization, which SimpleCov has known issues with.

This commit adds a few lines of extra config on parallelization setup/teardown to fix SimpleCov's naming of the individual coverage reports, and to correctly surface the coverage output from each worker