Add environment agnostic asset lookup
The previous approach here required assets:precompile to be run before
the code would work. This meant that when running the application tests
you needed to do the extra manual step of running assets:precompile,
otherwise there were failing tests - this was a common point of
confusion for developers as there isn't a clue that you needed to do
this.

Rather than adding assets:precompile as a step on the default rake task,
I've instead looked up how we can change the code to not be dependent on
this extra check.

An alternative, if we see future problems with this, would be to just do
a basic `File.exists?(Rails.root.join("app/assets/images",
fraction_asset_path))`.