Add rake task for republishing historically_political editions
We've recently done some work to expose a banner on the frontend for
hmtl attachments which were published alongside an edition which is
historically_political i.e. published under a previous government
and their opinion.

This rake task finds all documents where the editions are political and
historical with and have an associated html attachment and republishes
them.