Fix broken world location filter
The controller that processes this request accepts the world_locations
parameter as an array (reasons for this are unclear to me). It looks
like when this page was changed to the design system this functionality
was lost: https://github.com/alphagov/whitehall/commit/31ca46b03698bd27e87bb98d53e658cc92fce74d

This fixes the bug and adds tests that only pass with this change.