Replace complex document filter JS
In a bygone time Whitehall would provide finders for all sorts of
content with lots of filtering options. However now in 2022 it only
provides one basic finder - non-English content. And that only has a
single filtering field: world location.

This removes the old complex JS file, and the mechanism to use AJAX to
reload the page when the filter changes, with a very basic one that
submits the filter form if a user changes the lone select field.

This change was made to enable the removing of a shared_mustache as the
document_filter.js file was the last place in this application where
shared_mustache is used.