Replace last mustache template
With the change in a previous commit to remove the AJAX loading of
documents in a finder, this ports the mustache template to an erb file.

The erb does end up being a little strange, because the object that
generates the data, DocumentFilterPresenter [1], seems engineered for
use in mustache templates.

I decided against doing a wide refactoring as I'm aware that currently
work is going on that should lead to this page being removed from
Whitehall (famous last words).

[1]: https://github.com/alphagov/whitehall/blob/b6d26bb6e359f73821221c3a09a7cbf191837191/app/presenters/document_filter_presenter.rb