Only limit history for one specific document
This is a temporary short-term fix being applied to resolve an incident.

There is a known problem with one specific document that causes the Whitehall interface to timeout, due to the size of the document's version history.

The fix applied in 2346aa1829ab31c0471a63c3e1fbcad2e2a8a839 added a limit to the history of all documents. This solved the initial problem for that one document, but it's resulted in other documents having a limited history too.

This commit adapts the original fix so that it only applies to the affected document, identified by its Content ID. All other documents will have a regular, un-limited version history.