Refactor `Edition#submitted_by` method
This commit refactors the `Edition#submitted_by` method. This was done to improve both its readability and its performance/efficiency.

I've taken this opportunity to roll `Edition#most_recent_submission_audit_entry` in to `Edition#submitted_by`, since that was the only method using it anyway.

The original code in `Edition#most_recent_submission_audit_entry` was difficult to understand and not intention revealing. It also pulled out *every* Version associated with the Edition, and looped through in-memory to find the right version.

This new implementation leans on subqueries to offload work to the underlying MySQL database, meaning we only ever need to pull out one single Version and can avoid the loop altogether. Whilst it's still a complex query, I'm hoping the variable names alongside the comment should make this easier to understand for future developers.