Remove `Edition#publication_audit_entry`
This commit replaces the method `Edition#publication_audit_entry` with a new method `Document#first_published_on_govuk`.

The method was only used by the presenter method `DocumentListExportPresenter#first_published_on_govuk`, which used it to get the date that a document was first published on the GOV.UK website.

`Edition#publication_audit_entry` achieved this by looping through every Version of every Edition belonging to the Document, to find the first Version which had `state: "published"`. This represents the audit trail action which originally caused the document to be published on GOV.UK.

Instead, I've replaced it with the method `Document#first_published_on_govuk` which queries the database directly to find the first Version which moved the Document to a published state. This should be more efficient, especially when dealing with documents that have a large change history. It also felt more sensible for this method to exist directly on the Document model rather than the Edition.