SPIKE - proof of concept to only load document history rendered
This proof of concept explores a replacement to the document version
history lookup. The one that is replaced has the flaw that it will load
up every single history item regardless of how many are rendered (30).
This means that for a document with a long history there is a
significant risk of timeouts due to loading vast quantities of data.

We've looked particularly at a document produced by HMCTS, the Royal
Courts of Justice Cause list [1]. Which is updated multiple times a
day. This has led to a version history on this item to be > 9000 items.

The approach taken here is to try sidestep the complexity of
Edition::AuditTrail and implement similar objects but in a simpler way -
that start with a paginated collection.

I've put minimal thought into class names and organisation, so this
should be taken just a demonstration that it can be paginated.

[1]: https://www.gov.uk/government/publications/royal-courts-of-justice-cause-list