SPIKE - Demonstration of paginated editorial remarks
This is a somewhat incomplete demonstration of how we can make the
editorial remarks of an edition#show view use pagination.

There's a bunch of stuff it doesn't do like use JS to load pagination or
have a controller for that purpose.

It also would set-up some refactoring of the audit_trail bits for
document history as there's bits and pieces of logic for this in there,
and they could use a separate pagination parameter.

It also does something more than the current version does. The current
one breaks up the remarks into two blocks: those for current edition and
those for past editions. There's a bug with that if you look at a past
edition as it will indicate remarks on a newer edition is actually on a
past edition.

With the introduction of pagination, it's much harder to do a grouping
concept as you don't know if your page has all of the items of a
particular edition.

Instead I've entirely ordered the remarks chronologically (which seems
likely to be what everyone would expect) and as you traverse through
it'll add in headers for when things are on newer editions, current
edition or past editions.