Limit audit trail on documents with large history
This commit builds upon a 'quick fix' which was implemented in d06a2b41deca09b6d46e51e1b3ae80df788fd703 to avoid timeouts when publishers need to edit documents that have a large change history.

There is ongoing work to [make Whitehall better at handling large document histories][1] – but in the meantime, this is being put in place as a precautionary measure to avoid the potential for future incidents.

[1]: https://trello.com/c/SKFiAakd/493-handle-large-change-histories-more-efficiently-in-whitehall