Add a temporary shared_mustache:compile no-op task
The deployment process for Whitehall involves the calling of
`shared_mustache:compile` [1]. In order to maintain the ability to
deploy while removing shared_mustache, this commit adds a no-op task so
the deploy can succeed.

Once this code change has been deployed to production I shall update the
deployment script to remove the call to shared_mustache:compile and then
delete this task.

[1]: https://github.com/alphagov/govuk-app-deployment/blob/87dd69d98b0a77c1d7ed6bd7837800dda016fb74/whitehall/config/deploy.rb#L27-L34