Add rake task for bulk republishing missed html attachments
HTML Attachments can be associated with Editions or Responses. As part
of the work to add `auth_bypass_id`s for HTML Attachments we backfilled
them for all HTML Attachments which belong to an Edition.

However the backfill for HTML Attachments which are associated with
Responses were missed. These come in two shapes, both of which belong
to consultations.

1. `ConsultationOutcome`
2. `ConsultationPublicFeedback`

This commit adds a rake task which republishes documents which have a
draft HTML Attachment which belongs to an outcome or public feedback.
These always belong to a Consultation.