Remove unnused organisations CSS and related code
The organisations CSS is not used in the app (there are no uses of the
selector in the code). This removes it and related code that I could
find no trace of.

I probably got a little carried with the step definitions but these seem
very safe to remove as we'll get a failing test if I've been
over-zealous.