Add RepublishCsvAttachments rake task
Recently we've done some work to fix draft CSV rendering. As a
result the link to view the CSV attachment which is passed down for
the Frontend to render needs to be updated to contain the File's ID
rather then the AttachmenData ID.

Resultantly, we need to republish all documents with an associated CSV.
This commit adds a rake task to do this.