Update the CSV preview path to take the file id
At the moment, this endpoint is passed the attachment_data id. As
attachment data has many attachments for the same file, it is currently
not possible to distinguish which file has been requested from the ID
alone.

At the moment, the CSV preview endpoint defaults to rendering the
attachment belonging to the latest published edition for the attachment
data passed in.

We need to ensure that when a draft CSV is previewed it has the correct
styling and uses the draft editions details on the page.

This PR adds this functionality in by using the url to determine whether
a draft asset has been requested.# Please enter the commit message for
your changes.