Add Rake task codebase:metrics
For the last couple of months, we've been tracking some metrics
about the codebase, to help us check progress on our mission to
make it easier to work in.

So far, we've retrieved the metrics manually. This is error-prone,
and vulnerable to differences in method or OS shells, etc.

To mitigate this, I've added this Rake task that will execute as
many of the metrics as possible within the docker container.