Remove duplication from the controller
This pulls out quite a bit of duplication in the controller. It also
fixes a previous bug by ensuring that if the request is made by via
draft-assets, there is a draft edition. And if it is made via assets
then there is a visible (published) edition present.

Prior to this you could potentially have had draft assets render the
details of a visible edition.