Add an unpublished_at field to unpublishings
This change adds an "unpublished_at" field to the "unpublishings" table.

These Unpublishing records describe when and why an Edition of a Document was unpublished (including withdrawals).

Previously, the "created_at" timestamp was used to determine when an Edition was unpublished. However, we now have a need for users to be able to backdate withdrawals.

It makes sense at this point to introduce "unpublished_at" as a new database field of its own, and stop using the "created_at" timestamp as a proxy. It'll make it easier for us to set a backdated withdrawal date in the future without needing to override the "created_at" field. In doing so, we'll populate the existing records' "unpublished_at" fields to equal their respective "created_at" timestamps.