Add tests for Classification `next_ordering` method
There appears to be a bug where new topical event (aka classification)
features get incorrectly numbered when the list has previously been
sorted. The new feature always gets numbered as `1`, meaning it is in
the same position as another feature. The rendered ordering then varies
(maybe somewhat randomly).

The issue is coming from the `next_ordering` method, so adding tests to
determine this is the cause and allow us to refactor the method to work
correctly.