Update Classification ordering to start from zero
There is currently an anomaly in ordering of topical events (aka
classifications).

If you add a new feature to a new topical event, the ordering starts
from 1 and continues upwards. But if you reorder existing features on a
topical event, the ordering starts from 0 and continues upwards.

Updating the behaviour so both situations start from 0.