Fix Classification `next_ordering` method
This method was not previously working correctly, by incorrectly
ordering newly added features to existing feature lists  (see previous
commits where a test was added to demonstrate this).

The problem was caused as the `order` method did not override the
default scope, which orders the features in ascending order. Because of
this, the last item was actually the first and was ordered 0.

Changing to the `maximum` method means Rails will ignore the default
scope.