Fix 'Republish all documents with attachments' rake task
We've had a [report] that a particular publication was showing a
"Request an accessible format link" that doesn't do anything. This
is evidently a known issue and has been fixed already; what's
needed is to republish all affected documents.

The `repubish_docs_with_attachments` rake task was run, but did
not fix the publication specified in the report. On further
investigation, the publication has attachments which have a
`accessible` value of `nil`, not `false`. We therefore need to
widen the query to include documents containing attachments that
are both `accessible: false` and `accessible: nil`.

[report]: https://govuk.zendesk.com/agent/tickets/5017289