Rename all Classification tables to Topical Events
As Classification is only used by Topical Event, we are consolidating
the number of models and removing the single instance inheritance.

Because of this, we need to rename the relevant database tables to use
the new name.

Note that some of the index names had to be shortened to comply with
MySQL's 64-cahracter limit for index names. Unfortunately this means
Rails won't automatically rename the index names during a migration
involving the affected columns. This change is made manually here.