Add UI for backdating a document withdrawal
This commit adds to the existing 'withdraw or unpublish' journey so that users can choose to re-use an existing withdrawal date and explanation.

If a document has been withdrawn in the past, the user interface will display a list of previous withdrawal dates and allow users to choose one to re-use. Alternatively the user can choose to withdraw the document afresh using today's date and a new public explanation message.

This supports a user need where publishers may have withdrawn a document without updating attached files to mark them as 'withdrawn'. Publishers will now be able to un-withdraw the document, update the attachments accordingly, and then re-withdraw it using the previous withdrawal date. In this scenario, it would be misleading for the withdrawal date to change to the current date because it wouldn't be a truthful representation of when the document was actually withdrawn.