Add additional routes for Topical Event presentation
The atom feed for topical events is currently being rendered by
Whitehall, despite the code for this being migrated to Collections.

The reason being is router was not aware of the atom feed routes, so
used the default behaviour of sending all unregistered `/government`
requests to Whitehall.

Adding the additional route to the content that is presented to
Publishing API to resolve this issue.