Update failing tests
This fixes tests that are failing for what i would consider
`unrealistic` reasons. For example a document with a published
edition as the latest edition when a draft edition exists.

It also fixes a couple of tests which use FactoryBot#build then call
save! rather than FactoryBot#create, by manually updating
latest_edition_id and published_edition_id to use the correct values.