Ensure UpdateLiveAndLatestEditionId sets correct IDs based on state of objects
This updates the UpdateLiveAndLatestEditionId to set
live_edition_id and latest_edition_id based on the state
of the editions that belong to a document.

This is called after an edition is created in the factory and will
also be used when an edition is created, deleted or published to
ensure the they are remain correct when a documents editions transition
in state.