Add featured documents to world location news presenter
Also introduces a shared FeaturedDocumentsPresenter and uses this in organisations since the functionality is shared between the two content types. The limit for world location news featured documents is set to 5 as this is the maximum number we currently display on a page, whereas for organisations it is 6. We might want to re-consider this inconsistency in future. Setting the limit as a constant on WorldLocation will allow us to display a warning on the admin panel about the number of documents which we won't have to hard-code.