Update factories to populate latest_edition_id and live_edition_id
This does some preliminary work to try and ensure that latest_edition_id
and live_edition_id are populated correctly when objects
are instantiated by factories.

This is more proof of concept and I intend to create a service or
something similar which can be reused across the `EditionService` and
factories later which ensures latest_edition_id and live_edition_id
have the correct values.