Rename published_edition to live_edition
It doesn't really make sense that the published_edition association
is called that, as it can either be published or withdrawn. Instead it
makes more sense to rename it to live_edition to reflect that it is
live on gov.uk, but not necessarily published.

It then goes through and updates all the code and tests where it is
called.