Update the cucumber spec for editorial remarks
This now has a scenario which checks the notes tab is not
visible on the summary and edit page for editions, and the edit
page for translations.

Once the flag is pulled out this can be simplified and simply follow
the happy path with no checks for presence of the tabs.