Don't render notes tab for users with `View move tabs to endpoints`
This does the work to ensure that the Notes tab does not render on the
edition show, edition edit and translation edit pages.

I've passed an additional option into the
SidebarHelper#edition_tabs method which only adds the notes tab
when the user does not have the 'View move tabs to endpoints'
permission.

The last step is ensuring that notes and the link to adding a note
are removed from the sidebar.