Add index endpoint for editorial remarks (notes)
This adds an index endpoint where a user can go see view all the notes
for a document. This has an `Add note` link which sends a user through
the add note flow.

It ensures that when a user creates a new note they're redirected to the
editorial remarks index page if they have the
`View move tabs to endpoints` permission.