Ensure Document#editions are in order of creation
This change adds a sort order to the `Document#editions` association. This means we can always be confident a document's editions will be returned in the order they were created – with the oldest first, and the newest last.

Specifically, we sort by the `created_at` timestamp followed by `id` as a 'tie breaker', both in ascending order.

Previously the association didn't specify a sort order. It would have fallen back to MySQL's default behaviour, which would result in records tending to come back in the order they were created, but [that's not guaranteed][1]. In our case the sort order is important – because various areas of Whitehall implicitly depend on it being in order of creation – so it's not safe to rely on the default sort.

[1]: https://www.brentozar.com/archive/2020/04/why-ordering-isnt-guaranteed-without-an-order-by/