Add methods to get all attachments with changes on an edtion
- Adds a method to get deleted attachments
- Adds a method to gather all changed attachments

Created attachments are based on attachments created 15 seconds after
the edition is created. The time is used since attachments don't have a
concept of history and so cannot be compared to previous attachments.

30 seconds was selected since that is the amount of time a page takes to
time out. There are some editions which have so much history the page
times out during creation. If the time is set too low, some attachments
will be shown to be created erroneously.