Split World Location News out of World Locations
The `WorldLocation` model currently contains the data for two published
pages: the World Location itself and the World Location News page.

This has led to some unusual code needing to be written specifically for
the news page as publishing the `WorldLocation` model using
`PublishesToPublishingApi` and `Searchable` will lead to only the World
Location page being published.

By splitting the model into two parts, we will be able to remove this
custom code and re-use existing methods for publishing the news page.

We are using raw SQL in the migrations as these tables are translatable,
meaning Rails will return only some translations and not others if we
refer to the model by it's name. Using a `SELECT` statement will ensure
we get all translations.

Note that we are slightly changing the search behaviour here for when we
choose to send or not send world location news pages to search. We are
no longer sending them when their associated world location is inactive.
Here is an example of a news page for an inactive world location:
https://www.gov.uk/world/st-martin/news
which I don't think is worth sending to search (the rest are all similar).

Co-authored-by: Becka Lelew <becka.lelew@digital.cabinet-office.gov.uk>