Add PaginatedHistory class
This commit adds a new class in order to query the database in a paginated fashion.
This is to improve the speed at which data is returned in the case of excessively
large version histories, which were causing web pages to time out.
It also adds the associated tests for this class.