Update feature specs to use edition flow that will remain
Once we removed the `Redirect to summary page`, all users will no longer
see the `Save and continue` buttons.

At this point we would have to update all the feature specs to use the
new flow.

As at the moment, there are two ways to navigate through creating an
edition and tagging it, it seems easier to update the feature specs
to use the second flow, which will continue to exist after the
workflow rework.