Prevent change to updated_at when html_attachment is duplicated
When creating a new draft edtion the html_attachments are duplicated
to the new edition. It causes a re render of the govspeak content,
which causes the updated_at time to change despite the lack of change
to the govspeak_content. In addition the render_govspeak method is
performed asyncronously on a queue leading to wilding varying amounts of
time from the created_at time to the updated_at time.

As part of the work to display attachments with changes on new draft
editions it is crucial that we know the true updated_at time i.e. when
the govspeak_content is changed. Therefore we have stopped the updated_at
time from being changed when the draft edition is created.