Fix issue with translated items in world location news
This fixes a bug with the content items sent for world location news pages, whereby editing a translatable field, such as the mission statement, sent that field through in the content item for all locales. So this means that editing the French content for a world location news item would have resulted in French content appearing on the English page. The same would happen for editing the English content, with English content then appearing on the French page.

The fix in this commit is to switch the after_update hook to an after_commit hook for world location news pages. This addresses a known issue - see https://github.com/globalize/globalize/issues/659 - with the Globalize gem, whereby in an after_update or after_save hook does not respect the I18n locale that it is acting in the context of. After_commit does not have the same issue.

To note that we are adding the publishing to publishing api behaviour now when the world location page is saved, as well as when it is updated. Previously, this behaviour only happened on update. We have decided to introduce this behaviour now since when we switch to using the `PublishesToPublishingApi` trait, we will introduce this behaviour anyway. If this results in undesired behaviour (at worst, some empty world location / world location news pages published to the website), the fix for this should be to make these pages properly editionable and creatable via the UI, rather than continuing to hack around world location news pages.