Save latest_edition and live_edition on document
This change adds the method `Document#update_edition_references` which can be called on a document to update its 'latest edition' and 'live edition' references.

This is automatically called when editions are created and when they change state (e.g. from draft to published), ensuring that these references are kept up to date.