Remove expensive queries to get Document Editions
This commit changes the Document#latest_edition and Document#live_edition associations to use the recently introduced Document#latest_edition_id and Document#live_edition_id fields.

This is the final step outlined in PR #6733 where those fields were introduced.

The dynamic query previously used for `latest_edition` was particularly expensive. It'd been observed taking ~4 seconds to run on Documents with a large number of Editions. This new approach should be much quicker since no calculation needs to happen on the fly – instead, it's done at the point of saving Editions on Documents.