Remove Edition#latest_edition
The preferred way to get the latest Edition of a Document is to call `Document#latest_edition`.

For example:

```ruby
edition = Edition.last

edition.latest_edition

edition.document.latest_edition
```

This reflects the fact that Document is the source of truth for which Edition is currently the latest one. From a data modelling perspective, an Edition needn't know about other Editions that belong to the parent Document. This behaviour needn't be duplicated elsewhere, which is why I've removed it.