Scope nav sidebar changes to the editions interface
Recently, we made some changes to move History, Notes, and Fact checking
functionality from the sidebar to separate endpoints. However, we (I)
overlooked that there is a separate tab in Whitehall for `Corporate
information` which essentially reuses all the views from the `Editions`
tab.

This means that any changes we apply there are mirrored, which currently
is undesirable behaviour. We will apply them there as well later,
but for now it should retain the same behaviour.