Update world location news format for search api
This renames the format for world location news that is sent to rummager from "world_location_news_page" to "world_location_news". This is being done alongside work to migrate rendering for this content type away from whitehall. While the change in this commit is not required to move the rendering, it does fix a problem of consistency of naming which makes world location news tricky to deal with. So hopefully by making this change we simplify this content type for future work.