Increase the size of the custom jobs url field
Zendesk ticket: https://govuk.zendesk.com/agent/tickets/5054145

The user wants to add a link to the civil service jobs site to the
custom jobs url that has filters for the organisation.

The link they want to add is:
https://www.civilservicejobs.service.gov.uk/csr/index.cgi?SID=b3duZXJ0eXBlPWZhaXImcGFnZWNsYXNzPVNlYXJjaCZvd25lcj01MDcwMDAwJmNvbnRleHRpZD03MTE1NDI1JnBhZ2VhY3Rpb249c2VhcmNoY29udGV4dCZyZXFzaWc9MTY2MDkwNjUzMC1iNjY0ZjBlMWZiNDkyZmQ4OWRhZTg5YWFmYmI1ZjUxZmEyNWFlM2Jl

When they try to update the field, they get the following error:
`Mysql2::Error: Data too long for column 'custom_jobs_url' at row 1`

This is because the link has 258 characters, whereas the character limit
for a string field is 255. This is likely to be a problem for other
departments who want to also link to the Civil service jobs site.

Changing the type to "text" ups the character limit to 65,535.