Add task for republishing world location news and removing and re-adding to search
This will be needed to republish these document types since we have made / are making the following changes to world location news:

1) Adding more details into the content we send to publishing api to enable them to be rendered by a frontend app (not whitehall)
2) Changing the document type used in what we send to search.

Point 1) requires us to have a republishing task, and point 2) requires us to have a task that removes the old item from search (to avoid duplication of results) and re-adds the new one (with updated content type). I have included these in the same temporary task since we should ensure that we re-add as soon as possible after removing from the search index.

This task can likely be removed following the republishing of the documents with the new search format type.