Add task for republishing world location news without resending the pages to search
Since the world location news worker sends to both publishing api and search, I have changed its 'perform' signature to include a flag to say whether or not to publish to search, so that we can introduce a rake task that republishes without pushing to search. This is in line with how we republish other document types. It's likely that this will soon be unecessary, as we are planning a refactor to world location news pages to enable them to use the shared publishes_to_publishing_api and searchable modules, which will mean that this content type does not need a separate rake task for republishing. However, adding one as we will need it in the interim.