Change the signature of the should_protect_against_xss_and_content_attacks_on test helper
I'm committing this separately so that it's easier to see and in case someone disagrees with the way I've solved this issue.

Motivation: I want to specify a specific version of a factory that's used for this test, that stubs a method that's delegated to another class. Without this, I got into some gnarly infinite recursion with my factories. I also think that symbolising the test name to get the factory name is not a great pattern, I'd far rather explicitly see the factory that's being used. So this involved changing a lot of test files, but removing some slight magic which doesn't allow us to change the factory name.