Provide functionality to pass in class name to partial
At the moment, we use a lot of inheritance. This causes some issues
when we share form fields as they default to the parent class. For
example, attachments[title] for file attachments and html attachments.

We need to pass in the parent class of the object to link the errors to
the field inputs, but also need to retain the class of the object for
the title in the error summary component so we can state the class name.

For example, for a file attachment we need to pass in an Attachment
object to ensure errors link to the inputs and pass in FileAttachment so
that the error title can read:

`To save the file attachment please fix the following issues:`

This passes in an optional class_name variable which can be used to
construct the error summary title.