Don't flood Publishing API when ordering features
We noticed some undesirable behaviour which resulted in excessive calls being made to the Publishing API.

When re-ordering the 'features' (featured documents) associated with an organisation, we found that duplicate requests were being made to re-publish the organisation page – 'N' requests (where 'N' is the number of features on the organisation) instead of just 1.

This was happening as a result of the ActiveRecord 'after save' callback on the Feature model. When re-ordering, every Feature in a FeatureList gets an updated sort order – so each one gets saved in sequence, resulting in numerous executions of the 'after save' republish callback.

Thankfully there aren't many places in the code that create/save/delete Features, so I've moved the re-publish call out of the 'after save' callback and have moved it to the relevant places higher up in the code.